# openapi.yaml
openapi: 3.0.0
info:
  title: API Fizjoterapeutycznej Aplikacji
  version: 1.0.0
  description: Dokumentacja podstawowych endpointów z poprawionymi schematami na podstawie Mongoose models
servers:
  - url: http://localhost:8080
    description: Serwer lokalny
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Patient:
      type: object
      properties:
        _id:
          type: string
        personalInfo:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
            pesel:
              type: string
            dateOfBirth:
              type: string
              format: date
            gender:
              type: string
              enum: [M, F, Other]
            contact:
              type: object
              properties:
                phone:
                  type: string
                email:
                  type: string
                  format: email
        medicalInfo:
          type: object
          properties:
            allergies:
              type: array
              items: { type: string }
            chronicConditions:
              type: array
              items: { type: string }
            medications:
              type: array
              items:
                type: object
                properties:
                  name: { type: string }
                  dosage: { type: string }
                  frequency: { type: string }
            medicalHistory:
              type: array
              items: { type: string }
        physiotherapyData:
          type: object
          properties:
            referringPhysician: { type: string }
            initialDiagnosis: { type: string }
            treatmentGoals:
              type: array
              items: { type: string }
            contraindications:
              type: array
              items: { type: string }
            specialNotes: { type: string }
        consentGiven:
          type: boolean
        consentDate:
          type: string
          format: date-time
    Employee:
      type: object
      properties:
        _id:
          type: string
        personalInfo:
          type: object
          properties:
            firstName: { type: string }
            lastName: { type: string }
            contact:
              type: object
              properties:
                phone: { type: string }
                email: { type: string, format: email }
        roles:
          type: array
          items: { type: string }
        permissions:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [active, inactive]
        schedule:
          type: array
          items:
            type: object
            properties:
              date: { type: string, format: date }
              slots:
                type: array
                items:
                  type: object
                  properties:
                    startTime: { type: string }
                    endTime: { type: string }
    Appointment:
      type: object
      properties:
        _id: { type: string }
        patient:
          $ref: '#/components/schemas/Patient'
        employee:
          $ref: '#/components/schemas/Employee'
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled]
        notes:
          type: string
...
    Visit:
      type: object
      properties:
        _id: { type: string }
        patient:
          $ref: '#/components/schemas/Patient'
        employee:
          $ref: '#/components/schemas/Employee'
        appointment:
          $ref: '#/components/schemas/Appointment'
        date:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, in-progress, completed, cancelled]
        notes: { type: string }
        treatment:
          type: object
          additionalProperties: true
    Schedule:
      type: object
      properties:
        _id: { type: string }
        employee:
          $ref: '#/components/schemas/Employee'
        weekly:
          type: array
          items:
            type: object
            properties:
              day:
                type: string
                enum: [Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday]
              slots:
                type: array
                items:
                  type: object
                  properties:
                    startTime: { type: string }
                    endTime: { type: string }
    Notification:
      type: object
      properties:
        _id: { type: string }
        user:
          type: string
        type: { type: string }
        message: { type: string }
        read: { type: boolean }
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /api/patients:
    get:
      tags: [Patients]
      summary: Lista wszystkich pacjentów
      responses:
        '200':
          description: Lista pacjentów
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Patient'
    post:
      tags: [Patients]
      summary: Dodanie nowego pacjenta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Patient'
      responses:
        '201':
          description: Pacjent utworzony
...
  /api/appointments:
    get:
      tags: [Appointments]
      summary: Lista wszystkich terminów
      responses:
        '200':
          description: Lista terminów
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
    post:
      tags: [Appointments]
      summary: Utworzenie nowego terminu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Appointment'
      responses:
        '201':
          description: Termin utworzony
...
  /api/employees:
    get:
      tags: [Employees]
      summary: Lista wszystkich pracowników
      responses:
        '200':
          description: Lista pracowników
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
    post:
      tags: [Employees]
      summary: Dodanie nowego pracownika
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '201':
          description: Pracownik utworzony
...
  /api/visits:
    post:
      tags: [Visits]
      summary: Utworzenie nowej wizyty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Visit'
      responses:
        '201':
          description: Wizyta utworzona
...
  /api/schedules:
    get:
      tags: [Schedules]
      summary: Lista harmonogramów pracowników
      responses:
        '200':
          description: Lista harmonogramów
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'
    post:
      tags: [Schedules]
      summary: Dodanie nowego harmonogramu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
      responses:
        '201':
          description: Harmonogram utworzony
...
  /api/notifications:
    get:
      tags: [Notifications]
      summary: Lista powiadomień
      responses:
        '200':
          description: Lista powiadomień
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
    get:
      tags: [Notifications]
      summary: Szczegóły powiadomienia
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Dane powiadomienia
    post:
      tags: [Notifications]
      summary: Dodanie nowego powiadomienia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notification'
      responses:
        '201':
          description: Powiadomienie utworzone
    put:
      tags: [Notifications]
      summary: Edycja powiadomienia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notification'
      responses:
        '200':
          description: Powiadomienie zaktualizowane
    delete:
      tags: [Notifications]
      summary: Usunięcie powiadomienia
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Powiadomienie usunięte
...